moved classes into package store added InventoryTest.java
