+++ src/store/StoreManager.java
@@ -1,3 +1,5 @@
+package store;
+
 import java.util.ArrayList;
 
 /**
@@ -7,14 +9,14 @@ import java.util.ArrayList;
  * @version 1.1
  */
 public class StoreManager {
-    //init Inventory to some default values
+    //init store.Inventory to some default values
     private Inventory myInventory = new Inventory(new Product[]{new Product("apple", 1, 2.00), new Product("orange",2,2.5)}, new int[]{10,5});
     private ArrayList<ShoppingCart> userCarts = new ArrayList<>();
 
     /**
      * Get the store's inventory
      *
-     * @return Inventory, store's inventory
+     * @return store.Inventory, store's inventory
      */
     public Inventory getMyInventory() {
         return myInventory;
@@ -22,16 +24,16 @@ public class StoreManager {
 
     /**
      *
-     * @return ArrayList<ShoppingCart>, ArrayList of userCarts
+     * @return ArrayList<store.ShoppingCart>, ArrayList of userCarts
      */
     public ArrayList<ShoppingCart> getUserCarts() {
         return userCarts;
     }
 
     /**
-     * Check how much stock of a given Product is in the Inventory
+     * Check how much stock of a given store.Product is in the store.Inventory
      *
-     * @param myProduct Product, Product to check
+     * @param myProduct store.Product, store.Product to check
      * @return int, Returns amount of stock for given product. If product does not exist in inventory, returns -1.
      */
     public int checkStock(Product myProduct) {
@@ -53,7 +55,7 @@ public class StoreManager {
      * Removes stock from Store inventory and transfers stock to cart inventory
      *
      * @param cartID int, Cart ID of the desired cart
-     * @param productID int, Product ID of product to be added
+     * @param productID int, store.Product ID of product to be added
      * @param quantity int, amount of stock to add to the cart
      * @return return true if valid product ID's and stock levels are passed,
      *         else return false
@@ -73,7 +75,7 @@ public class StoreManager {
      * Removes stock from cart inventory and transfers stock to store inventory
      *
      * @param cartID int, Cart ID of the desired cart
-     * @param productID int, Product ID of product to be removed
+     * @param productID int, store.Product ID of product to be removed
      * @param quantity int, amount of removed from the cart
      * @return return true if valid product ID's and stock levels are passed,
      *         else return false
@@ -129,7 +131,7 @@ public class StoreManager {
     }
 
     /**
-     * Display the Store's Inventory
+     * Display the Store's store.Inventory
      */
     public void printInventory() {
         System.out.println("|--------------------THE COURSE STORE--------------------|");
