+++ src/store/StoreView.java
@@ -1,3 +1,5 @@
+package store;
+
 import java.util.Scanner;
 import java.util.ArrayList;
 
@@ -45,9 +47,9 @@ public class StoreView {
     }
 
     /**
-     * Add a Product to the User's cart
+     * Add a store.Product to the User's cart
      *
-     * @param id int, Product id
+     * @param id int, store.Product id
      * @param amount int, amount to add to user
      */
     public void addToUser(int id, int amount) {
@@ -57,9 +59,9 @@ public class StoreView {
     }
 
     /**
-     * Remove a Product from the user'sCart
+     * Remove a store.Product from the user'sCart
      *
-     * @param id int, Product id
+     * @param id int, store.Product id
      * @param amount int, amount to remove from user
      */
     public void removeFromUser(int id, int amount) {
@@ -102,19 +104,19 @@ public class StoreView {
         String s;
 
 
-        storeViews.add(new StoreView(sm, sm.newShoppingCart())); //add one existing StoreView
+        storeViews.add(new StoreView(sm, sm.newShoppingCart())); //add one existing store.StoreView
 
         while (!exit) {
             System.out.println("Please select a command:");
-            System.out.println("1: Select an existing StoreView");
-            System.out.println("2: Create a new StoreView");
+            System.out.println("1: Select an existing store.StoreView");
+            System.out.println("2: Create a new store.StoreView");
             System.out.println("-1: Exit \n");
 
 
             i = UserInput.getIntInput(-1, 2);
 
             if (i == 1) { //choose existing
-                System.out.println("Please select a StoreView or choose -1 to exit.");
+                System.out.println("Please select a store.StoreView or choose -1 to exit.");
                 System.out.println("Existing Store Views:");
                 UserInput.printArray(storeViews);
                 System.out.print("\n");
@@ -133,17 +135,17 @@ public class StoreView {
                             case "viewCart" -> storeViews.get(i).viewCart();
                             case "add" -> {
                                 storeViews.get(i).browse();
-                                System.out.println("Please enter the id of the Product you wish to add to your cart.");
+                                System.out.println("Please enter the id of the store.Product you wish to add to your cart.");
                                 id = UserInput.getIntInput();
-                                System.out.println("Please enter the amount of the Product you wish to add to your cart.");
+                                System.out.println("Please enter the amount of the store.Product you wish to add to your cart.");
                                 amount = UserInput.getIntInput();
                                 storeViews.get(i).addToUser(id, amount);
                             }
                             case "remove" -> {
                                 storeViews.get(i).viewCart();
-                                System.out.println("Please enter the id of the Product you wish to remove from your cart.");
+                                System.out.println("Please enter the id of the store.Product you wish to remove from your cart.");
                                 id = UserInput.getIntInput();
-                                System.out.println("Please enter the amount of the Product you wish to remove from your cart.");
+                                System.out.println("Please enter the amount of the store.Product you wish to remove from your cart.");
                                 amount = UserInput.getIntInput();
                                 storeViews.get(i).removeFromUser(id, amount);
                             }
@@ -157,7 +159,7 @@ public class StoreView {
 
             } else if (i == 2) { //add new
                 storeViews.add(new StoreView(sm, sm.newShoppingCart()));
-                System.out.println("New StoreView Created. \n");
+                System.out.println("New store.StoreView Created. \n");
             } else {
                 exit = true;
             }
