+++ src/storetest/InventoryTest.java
@@ -1,0 +1,37 @@
+package storetest;
+import org.junit.jupiter.api.Test;
+import store.Inventory;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class InventoryTest {
+
+    @Test
+    void getProductQuantity() {
+
+    }
+
+    @Test
+    void getProductInfo() {
+    }
+
+    @Test
+    void getStock() {
+    }
+
+    @Test
+    void addStock() {
+    }
+
+    @Test
+    void removeStock() {
+    }
+
+    @Test
+    void removeProduct() {
+    }
+
+    @Test
+    void getInfo() {
+    }
+}
